Moved onReciterViewed to PopularEntitiesProjector
